refactor: improve _unwrap_implicit_optional_hints impl
This PR changes the implementation so that it will flatten the optional union, if the other union arg is also a union.

E.g., `Union[Union[str, int], NoneType]` becomes `Union[str, int, Nonetype]`.

Adds extra test for missing condition where a function arg is annotated as a non-optional union with `None` default.