fix: typing of datastructures.URL (#2723)
This PR fixes an issue where mypy would infer the type of `datastructures.URL` to be `Any`. This was caused by the use of the `@lrucache` decorator on `URL.__new__()`. We had ignored the error, however mypy would report:

> litestar/datastructures/url.py:85: error: Unsupported decorated constructor type  [misc]

The fix adds a new `@classmethod`, `URL._new()` which is cached and called from `URL.__new__()`.