fix: connection cache access
It is currently possible to make handling of an HTTP request hang indefinitely by first creating two connection objects, and then accessing the connection data on each one. E.g., if we await `request_1.json()`, we'll receive the data, and then if we await `request_2.json()` it will not identify that the data has been cached and will indefinitely await on `receive()` for a message that will never arrive.

This PR resolves this issue by moving the check for cached data from the connection's `__init__()` method, to within the data access methods.

Closes #2727