Extend Litestar callable types and route handling.
The commit introduces OperationT in callable_types.py as a new TypeVar bound to the Operation class from openapi.spec.operation module. Following that, the OperationCallable was added to the list of type aliases.

For base.py in http_handlers, the operation_callable field was added with type 'OperationCallable | None' to the HTTPRouteHandler class and its constructor arguments. The value of this new field is later set in the constructor.

Lastly, in path_item.py, support for including the route based on the operation_callable runs and sets include_in_schema was added inside the build_path_item function.

These changes provide more flexibility and control in manipulating web application routes and their specifications in the OpenAPI format.