refactor: a new empty sentinel
WIP

The current `Empty` and `EmptyType` sentinel values and type are unable to be appropriately narrowed away, resulting in a lot of calls to `cast()`, particularly when accessing cached attributes on our core types.

This PR introduces a new empty sentinel, the adoption of which will allow us to properly narrow down from unions with itself.

This first stage implements the sentinels for our internal attribute caching only, and on transition to 3.0, `Empty` would alias `_EMPTY` and `EmptyType` would alias ``_LiteralEmpty`, so there shouldn't much downstream disruption.