refactor: change type of `Empty` and `EmptyType` (#2734)
* refactor: use scope state for caching connection data

WIP

* refactor: use type: ignore instead of cast() on hot path

* refactor: change type of `Empty` and `EmptyType`

This PR changes the type of `Empty` and `EmptyType` in `litestar.utils.empty` to types that better support narrowing.

There is no change to the API of empty, and identity checks against `Empty` will still work as they always have.

Use of `EmptyType` as a type in annotations will continue to work like it always has.

Type checking of `Empty` will improve, as evidenced by the bug that this PR exposed in our use of `ParameterDefinition.default_factory`. It is typed `Callable | None` and in multiple places we've been passing `Empty` to it which was passing type checks because the type `Empty` _is_ callable. As a single enum variant, this type of error should no longer be possible.

Where downstream users are importing and using `Empty` as one would expect, i.e., `if thing is Empty: ...`, this PR should only improve their typing experience.

An issue that may arise is if `Empty` is being used in an `issubclass` check, e.g., `issubclass(thing, Empty)`. Under this PR, this would now be a runtime error. However, given that the previous implementation of `Empty` was marked as `@final`, it cannot have been subclassed in user-code, and so I think safe to assume that this scenario would be rare, if not non-existent.

* refactor: improve exception wording

Fix wording of exception raised where a user might return the empty sentinel from a handler.