feat: stronger typing for scope state utils.
We use connection scope state to store a lot of different things for caching, auth, logging etc.

This PR improves the type safety of writing to, and retrieving from the scope state via the scope state utils.

The PR removes the "pop" behavior from `get_litestar_scope_state()` in favor of adding a new utility `pop_litestar_scope_state()` so that both the get and pop utils share the same semantics as `dict.get()` and `dict.pop()`.