fix: consistent union param errors
This PR fixes an error where unions of collection types would result in different errors depending on whether the union included `None` or not.

In either case, msgspec raises the same error, however the case where `None` was included in the union was resulting in an incorrect annotation being added to the signature model.

For #2600