fix: type narrowing in csrf middleware
This PR makes a small change to the CSRF middleware. By moving the test for `None` values outside of the `_csrf_tokens_match()` method we can appropriately narrow the type of `existing_csrf_token` removing the need for a `type: ignore`.

This also makes it more explicit that the tokens don't match when either are `None`.

Closes #2770