fix: raise config error when yield dependencies are cached
This PR makes it a configuration error to attempt to `use_cache=True` with generator dependencies.

Prior to this implementation, this would result in a 500 error at runtime.

The reason for making this a config error is that after the first use of the generator the cleanup code executes and so subsequent use of the generated value is dubious.

Closes #2771