fix: deduplication of openapi params
Previous behavior would consider two parameters with the same name but declared in different places (eg., header, cookie) as an error.

This PR incorporates the "param_in" value when validating params for openapi spec so that it would only be an error to have multiple different parameters, of the same name declared in the same place.

Closes #2662