fix: recursive schema generation
This PR modifies the way we generate schemas for objects that should be able to be referenced by other schemas.

It adds a new method to `SchemaCreator` called `create_component_schema` that should be used to create a schema that should be able to be referenced by other schemas, and itself.

This method takes care to first create the schema in the registry so that it can be immediately referenced.

In `SchemaCreator.for_field_definition()`, before we generate a schema for any plugin-supported type, we first check whether a reference can be made to an already existing schema and return that if possible.

Closes #2429