fix: correctly parse files when uploading list of files
When a model is specified for uploading files and a specific field in that model is annotated with
`list[UploadFile]` and a single file is uploaded, the validation fails. This is because the extracted
file is not passed onto `msgspec` as a list. This fixes that by ensuring that the value is converted
into a list if needed before giving it to `msgspec` for validation.