fix: sse event type is only sent once (#2888)
* Refactor ServerSentEvent to ServerSentEventStream

The `ServerSentEvent` class has been renamed to `ServerSentEventStream` across the entire codebase including tests and documentation. This change was made to provide a more accurate name for the class and reflect its purpose better.

* Refactor class in sse.py

The class `_ServerSentEventIterator` in `sse.py` has been refactored. Features include a new `chunks` attribute and a minor modification to the `_call_next` method.

---------

Co-authored-by: Janek Nouvertné <provinzkraut@posteo.de>