fix: cookie naming for short cookies (#3095)
* Fix cookie naming for short cookies

Previously, cookie names always had a suffix of the form "-{i}"
appended to them. With this change, the suffix is omitted if the
cookie is short enough (< 4 KB) to not be split into multiple chunks.

If the cookie is so big that it is split into several chunks, the
chunk number suffix is appended, as before.