refactor: openapi router (#2893)
* refactor: openapi router

This PR refactors the way that we support multiple UIs for OpenAPI.

We add `litestar.openapi.plugins` where `OpenAPIRenderPlugin` is defined, and implementations of that plugin for the frameworks we currently support.

We add `OpenAPIConfig.render_plugins` config option, where a user can explicitly declare a set of plugins for UIs they wish to support.

If a user declares a sub-class of `OpenAPIController` at `OpenAPIConfig.openapi_controller`, then existing behavior is preserved exactly.

However, if no controller is explicitly declared, we invoke the new router-based approach, which should behave identically to the controller based approach (i.e., respect `enabled_endpoints` and `root_schema_site`).

Closes #2541

* docs: start of documentation re-write.

- creates an indexed directory for openapi
- removes the controller docs
- start of docs for plugins

* refactor: move JsonRenderPlugin into private ns

We add the json plugin, and have hardcoded refs to the path that it serves, so best not to make this public API just yet.

* docs: reference docs for plugins

* Revert "refactor: move JsonRenderPlugin into private ns"

This reverts commit 60719aabd31255ff315eb6dda768478dee371b8f.

* docs: JsonRenderPlugin undocumented.

* docs: continue plugin docs

* test: run tests for both plugin and controller

Modifies tests where appropriate to run on both the plugin-based approach and the controller based approach.

* Implement default endpoint selection logic.

* Deprecation of OpenAPIController configs

* docs: swagger oauth examples

* Update docs/usage/openapi/ui_plugins.rst

* Update docs/usage/openapi/ui_plugins.rst

* fix: linting

* refactor: don't rely on DI for openapi schema in plugin handler.

* fix(test): there's an extra schema route to serve 404s.

* fix(docs): docstring indent

* fix(lint): remove redundant return

* refactor: plugins receive style tag instead of tag content.

* feat: allow openapi router to be handed to openapi config.

Allows for customization, such as adding guards, middleware, other routes, etc.

* feat: add `scalar` schema ui (#2906)

* Update litestar/openapi/plugins.py

Co-authored-by: Jacob Coffee <jacob@z7x.org>

* Update litestar/openapi/plugins.py

Co-authored-by: Jacob Coffee <jacob@z7x.org>

* Update litestar/openapi/plugins.py

Co-authored-by: Jacob Coffee <jacob@z7x.org>

* Update litestar/openapi/config.py

Co-authored-by: Jacob Coffee <jacob@z7x.org>

* Update litestar/openapi/config.py

Co-authored-by: Jacob Coffee <jacob@z7x.org>

* fix: update deprecation version

* fix: use GH repo for scalar links

* fix: update default scalar version

* fix: scalar plugin style attribute render.

Plugins expect that the style value is already wrapped in `<style>` tags.

* fix: serve default static files via jsdeliver

* fix: docstring syntax

* fix: removes custom repr

Can always add if there's a need for it, but we aren't using it.

* docs: another pass

* fix: style

* fix: test for updated build openapi plugin example

* fix: absolute paths for openapi.json

Resolves #3047 for the openapi router case.

* refactor: simplify scalar plugin.

* fix: linting

* Update litestar/_openapi/plugin.py

* refactor: test app to use scalar ui plugin

* fix: plugin customization example

Version arg is ignored if `js_url` is provided.

* fix: remove unnecessary kwargs

Removes passing default values to plugin kwargs in examples.

* fix: grammar error

* feat: make OpenAPIRenderPlugin an ABC

Abstract on `render()` method.

* fix: correct referenced lines

Referenced LoC in example had drifted.

* fix: more small docs corrections

* chore: remove dup spec of enabled endpoints.

* fix: simplify test.

---------

Co-authored-by: Jacob Coffee <jacob@z7x.org>