fix(openapi): set default on schema from `FieldDefinition` (#3280)
Consider the following:

```python
    def get_foo(foo_id: int = 10) -> None:
        ...
```
In such cases, no `KwargDefinition` is created since there is no
metadata provided via `Annotated`. The default is still parsed, and
set on the generated `FieldDefinition`, however the `SchemaCreator`
currently only considers defaults that are set on `KwargDefinition`.

So in such cases, we should fallback to the default set on the
`FieldDefinition` if there is a valid default value.