feat: Add `path` parameter to Litestar application class (#3314)
* feat: allow for console output to be silenced (#3180)

In some cases, I've wanted to change the name of the "Application" to show something else instead of "Litestar".  This is usually to make a CLI feel more cohesive and part of a larger application.  In the same vein, I've found cases where I wanted to complete suppress the initial `from_env` or the Rich info table at startup.