refactor: middleware optimizations
Only wrap `route.handle` if there is middleware. This is because middleware may wrap the `send()` coroutine, and that send wrapper needs an exception handler middleware between itself and the handler in order for the wrapper to be called if the handler raises an exception.

Only wrap the middleware stack if there are exception handlers defined for the route. If there are no