refactor: exception handler middleware
Store the resolved exception handlers for the route on the connection scope, and retrieve them from within the exception middleware.

This removes the requirement for the exception handler middleware that wraps the resolved middleware stack as the outer most exception middleware can have access to the exception handlers that have been determined after route resolution.