Improve few documentation pages
The main goal was to update code examples in tutorial to include imports
as well, otherwise it can be quite annoying to scroll between text and
full code to see from where a specific class or function is imported if
one is writing code from scratch following the tutorial.

Other part of changes are related to minor typo, grammar, word
placement, formatting (both text and code) and similar fixes.