fix: register new CLI commands with `rich-click` (#4298)
After the most recent updates to `rich-click` and `click`, CLI plugins no longer show the command in the command help text.

This is because the plugin was not loaded before rending the help text.  This change ensures that the plugins are loaded prior to formatting the help output.

(cherry picked from commit 980029f784d3bffd6dcca7e1a64e3ac584923f9d)